update graphql qty type to be float

+++ app/graphql/types/item_type.rb
@@ -1,7 +1,7 @@
 module Types
   class ItemType < Types::BaseObject
     field :id, ID, null: false
-    field :qty, Integer
+    field :qty, Float
     field :name, String
     field :unit_type, String
     field :price, Float
